cloud/requirements.txt
kuksa-client
eclipse-zenoh
python-dotenv
